backend/requirements.txt
Flask
Flask-Cors
google-generativeai
python-dotenv
Pillow
openpyxl
gunicorn